web/backend/requirement.txt
absl-py==0.9.0
asgiref==3.2.7
asn1crypto==1.3.0
astor==0.8.0
beautifulsoup4==4.9.0
blinker==1.4
cachetools==3.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cycler==0.10.0
decorator==4.4.2
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
drf-yasg==1.17.1
futures==3.1.1
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.27.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
imgaug==0.4.0
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mysqlclient==1.4.6
networkx==2.4
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.1.0
packaging==20.3
Pillow==7.0.0
protobuf==3.11.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pydot==1.4.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
tornado==6.0.4
tqdm==4.45.0
uritemplate==3.0.1
urllib3==1.25.9
Werkzeug==0.16.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
